Implementación de migraciones para las tablas case_types y legal_cases. Se ha creado la migración para la tabla case_types, que incluye campos para nombre y descripción, y se ha implementado la migración para legal_cases, que incluye campos como código, fechas y una relación con case_types. Además, se han realizado ajustes para evitar la eliminación de tablas en las migraciones, mejorando la gestión de la base de datos.

+++ database/migrations/2025_05_15_230406_create_legal_cases_table.php
@@ -11,16 +11,18 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('legal_cases', function (Blueprint $table) {
-            $table->id();
-            $table->string('code', 50)->unique();
-            $table->date('entry_date');
-            $table->foreignId('case_type_id')->constrained()->onDelete('cascade');
-            $table->date('sentence_date')->nullable();
-            $table->date('closing_date')->nullable();
-            $table->timestamps();
-            $table->softDeletes();
-        });
+        if (!Schema::hasTable('legal_cases')) {
+            Schema::create('legal_cases', function (Blueprint $table) {
+                $table->id();
+                $table->string('code', 50)->unique();
+                $table->date('entry_date');
+                $table->foreignId('case_type_id')->constrained()->onDelete('cascade');
+                $table->date('sentence_date')->nullable();
+                $table->date('closing_date')->nullable();
+                $table->timestamps();
+                $table->softDeletes();
+            });
+        }
     }
 
     /**
@@ -28,6 +30,6 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('legal_cases');
+        // No eliminamos la tabla aquí para evitar conflictos con otras migraciones
     }
 };
